Create separate files for Hilt modules
